+++ lib/paper_trail/version_number.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module PaperTrail
   # The version number of the paper_trail gem. Not to be confused with
   # `PaperTrail::Version`. Ruby constants are case-sensitive, apparently,
@@ -5,9 +7,9 @@ module PaperTrail
   # because of this confusion, but it's not worth the breaking change.
   # People are encouraged to use `PaperTrail.gem_version` instead.
   module VERSION
-    MAJOR = 7
-    MINOR = 0
-    TINY = 3
+    MAJOR = 8
+    MINOR = 1
+    TINY = 1
     PRE = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".").freeze