+++ spec/models/callback_modifier_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "spec_helper"
 
 RSpec.describe CallbackModifier, type: :model, versioning: true do
@@ -15,12 +17,17 @@ RSpec.describe CallbackModifier, type: :model, versioning: true do
       end
     end
 
-    context "when :after" do
-      it "creates the version after destroy" do
-        modifier = AfterDestroyModifier.create!(some_content: FFaker::Lorem.sentence)
-        modifier.test_destroy
-        expect(modifier.versions.last.reify).to be_flagged_deleted
+    if ActiveRecord.gem_version < Gem::Version.new("5") ||
+        !ActiveRecord::Base.belongs_to_required_by_default
+
+      context "when :after" do
+        it "creates the version after destroy" do
+          modifier = AfterDestroyModifier.create!(some_content: FFaker::Lorem.sentence)
+          modifier.test_destroy
+          expect(modifier.versions.last.reify).to be_flagged_deleted
+        end
       end
+
     end
 
     context "when no argument" do