+++ spec/spec_helper.rb
@@ -1,8 +1,9 @@
-require "pry"
+# frozen_string_literal: true
 
-# This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV["RAILS_ENV"] ||= "test"
 ENV["DB"] ||= "sqlite"
+
+require "byebug"
 
 unless File.exist?(File.expand_path("dummy_app/config/database.yml", __dir__))
   warn "WARNING: No database.yml detected for the dummy app, please run `rake prepare` first"
@@ -72,7 +73,7 @@ RSpec.configure do |config|
   # Truncation is about three times slower than transaction rollback, so it'll
   # be nice when we can drop support for rails < 5.
   if active_record_gem_version < ::Gem::Version.new("5")
-    config.before(:each) { DatabaseCleaner.start }
-    config.after(:each) { DatabaseCleaner.clean }
+    config.before { DatabaseCleaner.start }
+    config.after { DatabaseCleaner.clean }
   end
 end